Refactor: Create Navigation Logic
