+++ app/src/main/AndroidManifest.xml
@@ -57,13 +57,7 @@
             android:name=".CreateProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".TestActivity"
-            android:exported="false" />
-        <activity
             android:name=".test.ui.login.LoginActivity"
-            android:exported="false" />
-        <activity
-            android:name=".StartActivity"
             android:exported="false" />
         <activity
             android:name=".SplashActivity"
